Calculate MINE statistics.

Calculate MINE statistics.

tidymine(
  dataset,
  outcome,
  predictors = -{     {         outcome     } },
  alpha = 0.6,
  c = 15,
  cores = getOption("mc.cores", 1L),
  low_variance_threshold = 1e-05,
  eps = NULL,
  est = "mic_approx",
  na.rm = FALSE,
  use = "pairwise.complete.obs",
  normalisation = FALSE,
  ...
)

Arguments

dataset
The dataset containing the outcome and predictors.

outcome
The variable of interest.

predictors
The variables to measure the predictive power of. Defaults to every variable apart from the outcome.

cores
Number of cores to use when executing the chains in parallel. This defaults to 1 (i.e. not performing parallel computing). Running `options(mc.cores = parallel::detectCores())` will set this be the maximum number of processors as the hardware and RAM allow (up to the number of chains).

low_variance_threshold
A parameter to remove predictors with low variance. MINE statistics cannot be calculated for variables With low variance. This is fine because a low variance predictor cannot have high predictive power.

est
Default value is "mic_approx". With est="mic_approx" the original MINE statis-tics will be computed, with est="mic_e" the equicharacteristic matrix is is evaluated and the mic() and tic() methods will return MIC_e and TIC_e values respectively.

na.rm
Logical to remove NA values.

use
Method to deal with NA values.

...

Value

MINE statistics of predictors for predicting outcome.

Examples

spellman %>%
  tidymine(x40)

#> Error in spellman %>% tidymine(x40): could not find function "%>%"

who %>%
  tidymine(lung_cancer_deaths_per_100_000_women)

#> Error in who %>% tidymine(lung_cancer_deaths_per_100_000_women): could not find function "%>%"